oracle/requirements.txt
Faker
cx_Oracle
python-decouple
python-dotenv
